+++ group-buy-trade-platform-app/src/test/java/com/tinuvile/test/domain/trade/ITradeLockOrderServiceTest.java
@@ -10,7 +10,7 @@ import com.tinuvile.domain.trade.model.entity.MarketPayOrderEntity;
 import com.tinuvile.domain.trade.model.entity.PayActivityEntity;
 import com.tinuvile.domain.trade.model.entity.PayDiscountEntity;
 import com.tinuvile.domain.trade.model.entity.UserEntity;
-import com.tinuvile.domain.trade.service.ITradeOrderService;
+import com.tinuvile.domain.trade.service.ITradeLockOrderService;
 import lombok.extern.slf4j.Slf4j;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -27,13 +27,13 @@ import javax.annotation.Resource;
 @Slf4j
 @RunWith(SpringRunner.class)
 @SpringBootTest
-public class ITradeOrderServiceTest {
+public class ITradeLockOrderServiceTest {
 
     @Resource
     private IIndexGroupBuyMarketService indexGroupBuyMarketService;
 
     @Resource
-    private ITradeOrderService tradeOrderService;
+    private ITradeLockOrderService tradeOrderService;
 
     @Test
     public void test_lockMarketPayOrder() throws Exception {
